requirements.txt
# requirements.txt

# core & environment
Flask
python-dotenv
requests

# async & background tasks
aiohttp
asyncio

# data
pandas
jsonschema

# dogging & errors
rich
loguru

# api integration
openai
httpx

# n8n & webhook connection
fastapi
uvicorn
gunicorn

# telegram & twitter
python-telegram-bot
tweepy
